[shell] Propagate key events from EvWindow to EvView
EvAnnotationWindow is a GtkWindow that doesn't accept focus, but we
manually grab the focus on the text view contained in it to be able to
modify the annotation contents. We need to propagate the key events to
the view even when the view is not the focused widget. The view will
only handle these events when there's a popup window with the focus so
that it's possible to type in window annotations.