Remove zoom-invalid signal from EvView
It was used by EvView when sizing mode was best-fit or fit-width to
request the window for its content size. This is not needed at all,
since we already call size_allocate on parent class. GtkScrolledWindow
takes into account the shadow type and scrollbars visibility in its
size_allocate method, so we can just take the allocation returned
to calculate the zoom in case of best-fit or fit-width.