Use a dynamic pixbuf cache size based on document page size
Instead of using a static number of pages to cache, we use a size in
bytes, and the number of pages that will be cached depends on the
current zoom level. It allows us caching more pages for lower scale
factors and increase zoom level by caching fewer pages. See bug #303365.