Do not remove unexistent files from recent menu
The problem is that gtk_recent_info_exists() does a stat on the file to
check whether it exists, and thta causes a problem on locally mounted
remote or external file systems. So, I think it's better to have
unexitent files in the recent menu (which is unlikley) than slowing down
evince startup.

https://bugzilla.gnome.org/show_bug.cgi?id=674188