ev-page-action-widget: Use g_signal_connect instead of g_signal_connect_object
g_signal_connect_object has been finally fixed in recent versions of
glib, so the signal is disconnected when the object is disposed. However
we are manually disconnecting the signal handler, which gives a runtime
warning with recent glib versions because the signal has already been
disconnected. Use g_signal_connect() and keep disocnnecting the signal
manually to make sure it still works with previous versions of glib.