ev-view: call gtk_style_context_set_background()
Even if we now render a background in _draw() with
gtk_render_background, using gtk_style_context_set_background() in
realize makes the backing GdkWindow not being completely transparent,
which enables GTK to apply optimizations during the paint cycle.
The result is, especially in clutter-gtk applications such as Documents,
scrolling in a document page is much smoother.

https://bugzilla.gnome.org/show_bug.cgi?id=692988