shell: Show as much text as possible for index tree
Uses separated columns for index and page number in the
sidebar for the table of contents (aka index).  This makes
to use as much space as possible for the index column with
a reasonable size for the page number for the most common
cases (e.g. page number < 9999 or 'cover', 'front', 'back').
For longer page numbers it uses ellipsis.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=676593

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>