shell: increase search performance. Fixes #667569.
Reduces text search duration by reducing the update frequency.

For documents with 100 pages or less, the GUI is updated for
every page.  For documents between 101 and 200 pages, the GUI
will be updated every other page, between 201 and 300, the GUI
will be updated every 3 pages, and so on.

In the PDF reference 1.7 (1310 pages), searching a word like
Quadrilat, the number of status updates in the GUI went down
from 1310 to 93 times, that also reduced the time taken to show
the first match (in page 612) from ~11 secs to ~6 secs.