backends: Don't ask for bug report in known issues. Fixes #695165.
When an annotation is unimplemented, Evince warns the user and
asks to report the issue here in Bugzilla with a test case.
However, there are several reports with test cases that have
not been implemented yet, and for ones we do not need more
reports (they would be marked as duplicated anyway).

The patch add a serie of known unimplemented annotations, warns
the user but does not ask her or him to file a bug.