ev-window: do not remove find job when find bar closes
In order to have the find next and previous keyboard shortcuts function even
when the find bar is closed we should not clear and cancel the find job in
ev_window_close_find_bar. The job still gets cleared and canceled on a new
search which happens every time the search string changes.

And for that reason do not call ev_window_search_start when the find bar goes
visible, as that will clear and cancel present find job hindering the
next and previous operations.

We also need to run the ev_find_{next|previous} functions as idles in order
to make sure that the view has time to update when the window was hidden.

https://bugzilla.gnome.org/show_bug.cgi?id=625225