ev-utils: improve Save Image As dialog
This commit makes the dialog more like the 'Open...' dialog.
The 'By Extension' filter is changed to an 'Supported Image Files' filter with all the
supported mime types added. An 'All Files' filter is added as well.

And to make it more like other dialogs, the supported formats are added with
just the description as name.

https://bugzilla.gnome.org/show_bug.cgi?id=523071