Check if find bar is visible when ESC is pressed to hide it
We were checking if the focused widget was a children of the find bar.
Now that there's a find sidebar and findbar is not closed when it loses
the focus, other widgets might have the focus while a search is active.
So, simply check if the findbar is visible instead.