ev-find-sidebar: Show all results in sidebar for multicolumn documents
Search results are sorted in physical order while text layout is in
reading order, so when a match is not found after the offset of previous
one, continue iterating the text layout from the beginning. Add a
warning when we fail to find the offset for a match because it should
never happen.