debug: Add support for drawing different debug borders in EvView
When evince is built with debug support, the env variable
EV_DEBUG_SHOW_BORDERS can be used to draw different debug borders in
EvView using different colors. Depending on the value of the env
variable and and the features supported by the current backend, it can
draw bounding boxes of characters, links, images, form fields,
annotations and selection region.