Use a GtkEntry with an arrow icon instead of a GtkComboBox for zoom selector
This allows us to use our own custom menu with toggle and non-toggle
menu items. Fit-page and fit-width options are toggle menu items, so
that it's possible to know which mode is currently in use. The code is
simpler now too and zoom selector actions are handled like all other
chrome actions.