Keep track of the current search result page instead of using the current page
In continous mode the page of the current search result might not the
the current page. Also make sure that when the search is restarted (the
find bar is closed and opened with the same search string) both the view
and the sidebar highlight the first match found from the current page.

https://bugzilla.gnome.org/show_bug.cgi?id=710778