Change the condition for restarting search after reloading
The bug is because the relevant check condition gets the focused
widget and it's ancestor. The problem lies in the fact that if the
user takes away focus from the sidebar, ancestor is no longer the
'Find' sidebar and the condition evaluates to FALSE even though it
shouldn't. Changing the condition to check if the 'Find' sidebar is
visible solves the problem by restarting search only when the 'Find'
sidebar is visible.

https://bugzilla.gnome.org/show_bug.cgi?id=703595