printing: Fix centering when shrinking to printable area but not scaling
When shrinking to printable area is selected and we don't need to scale
because the document fits in the printable area, the document is
centered applying the current scale. Since we are not going to scale in
that case, we should ignore any scale value > 1 when shrinking to
printable area.