appdata: Do not use the generic name in the software center
In GNOME3 Evince is a core module, so installed by default. We want to show the
project name in the software center so that it can be installed on other
desktops. This also means we don't get multiple "Document Viewer" items in the
search results that correspond to different projects.