find-sidebar: Do not use gtk_list_store_clear()
Set a new model instead. It seems to be more efficient because
gtk_list_store_clear() ends up emitting row-deleted signal for every row
in the model. We don't want the tree view to relayout, and update
according to the model for every row since we are going to remove all of
them.