print-operation: Fix centering of documents when printing with a manual scale
When a manual scale is used for printing, the document is not correctly
centered because the cairo context has a scale already applied. We need
to consider the current scale when centering the page and also convert
the coordinates before the cairo_translate().
Additional fixes to the offsets calculations by Leo Wolf <jclw@ymail.com>.

https://bugzilla.gnome.org/show_bug.cgi?id=734788