Fix runtime critical warning when starting in fullscreen mode
The problem is that when running fullscreen mode we call
ev_window_update_links_model() to update the page action widget of the
fullscreen toolbar, but when running directly in fullscreen mode, this
can happen before the links have been loaded in the sidebar. Since
ev_window_update_links_model() is also called automatically when the
links model property changes, we can simply return early in
ev_window_update_links_model() when the model is NULL and it will be
called again with a valid model when the links job finishes.

https://bugzilla.gnome.org/show_bug.cgi?id=737864