Add shortcuts for Continuous and Dual modes
The "Continuous" and "Dual" menu entries are among the few that do not have a shortcut.
This patch adds the shortcut 'c' to toggle the Continuous mode, and the sortcut 'd' to toggle the Dual mode.
Together with the existing shortcuts 'f' and 'w', they allow to quickly change between different layouts.

https://bugzilla.gnome.org/show_bug.cgi?id=670964