Show correct page when next search result requested
If there are more than 1 results on a page then requesting
next search result from the same page will not return you
to the page with the search results if you moved from the
page in the meantime (when you are not in continuous mode).
This applies also to previous results.

https://bugzilla.gnome.org/show_bug.cgi?id=730252