shell: Never save n-copies in the print-settings file
By default gtk_print_settings_to_key_file save all settings. When fixing
https://bugzilla.gnome.org/show_bug.cgi?id=696203 we re-introduced
https://bugzilla.gnome.org/show_bug.cgi?id=488806 for the number of copies.
This happens because the code that saves the document print settings in the evince
metadata deletes the document print settings from the key file. Thus,
the fix for bug #696203 is not stricty correct because we still
need to delete n-copies from the key_file. This patch fixes that.

https://bugzilla.gnome.org/show_bug.cgi?id=748549