ev-properties-view: Add accessible relations between the labels
The properties view grid consists of two columns, the left column
contains a GtkLabel which describes, or "labels," the property
displayed in the GtkLabel in the right column. The lack of accessible
relations between the labels in each column means Orca has no clue
that it should also present the label on the left when the user
Tabs amongst the labels on the right, and that it should not treat
labels on the left as static text.

https://bugzilla.gnome.org/show_bug.cgi?id=749127