ev-media: Make EvMediaPlayer a self contained widget to show the video and the controls
Using the EvView window with gst_video_overlay_set_render_rectangle()
to paint the videos seems to cause issues in some cases where the video
is not correctly positioned, or uses the whole view area. To avoid that,
we use now a specific widget (a GtkDrawingArea) to render the video
instead. EvMediaPlayer is not a GtkBox containing the drawing area and
the media controls. This simplifies the code a lot and requires less
code in EvView which will also make easier to add support for videos
in presentation view.