Fix make distcheck
It's seems this is a bug in intltool that happens since automake 1.15,
that uses a subdir sub inside the buildir during distcheck. I have no
idea how to fix it, so this is a temporary workaround so that I can make
releases until the bug is fixed.

See https://bugs.launchpad.net/intltool/+bug/1117944