libview: Increase page duration resolution
In presentation mode, if a page duration is specified, then the viewer
should advance to the next page. The current time resolution is
multiples of 1 second, which is too coarse for animations where each
page lasts for a fraction of a second.

By using g_timeout_add we increase the duration resolution, as
duration is a double, it can contain fractions of seconds, and
takes the advantage that g_timeout_add resolution is miliseconds.

Fixes #637

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>