Check legal boundaries of accessible pages
When a document is reloaded, the accessible->priv->children is
cleaned up and updated with the new document information.
In specific cases, where the document is smaller, priv->end_page
will be higher than the current number of pages. Therefore, we
check the number of pages to avoid access to nonexistent pages
in priv->children).

Fix https://bugzilla.gnome.org/show_bug.cgi?id=735744