Fixed the calculus about the maximum width of the label with the page count
With this patch Evince doesn't keep in memory the page_labels array
anymore for the documents that have the text of the page label equal to
the page number. In this manner is possible to manage properly the
documents that have a custom page label calculating the right width for
the label with the page count.

https://bugzilla.gnome.org/show_bug.cgi?id=764771