window: don't use a thumbnail as window icon
Modern desktop environments already have window previews e.g. in the
Alt-Tab interface (or in the Activities overview, in case of
gnome-shell).
For desktop environments that don't support an app menu, this also has
the effect of making the "application icon" used for the menu that
appears in the top-left window corner be the thumbnail, which is
confusing and unclear.

Just remove the code that changes the window icon.

https://bugzilla.gnome.org/show_bug.cgi?id=767747