libview: Do not create annotation window twice.
Currently, in button_release_event we always create the annotation window. Since
there is a chance that draw event will show the annotation windows (and in that
case creates annotation windows if they don't exist) then the annotation window
will be created twice. This causes a leak because we only destroy one of these windows...
and also when removing annotations with the window open we don't destroy the window
(we destroy a window which is already hidden).

The solution here is to check in button_release wether a window has been already created
for the annotation, and only if not, create the new annotation window.

https://bugzilla.gnome.org/show_bug.cgi?id=769200