libview: only access the relevant page cache for the height request
In ev_view_get_height_to_page, the "height" or "dual_height"
pointer can be set.  If dual_height is set, the dual_height buffer
is used. Using the "normal" height cache in dual view causes
an out of bounds read. So we restrict the cache accesses to their
relevant buffers.

https://bugzilla.gnome.org/show_bug.cgi?id=771612