attachments: Save attachments in a temporary directory
In order to present the file name "as is", it makes evince store
its temporary attachment inside a `evince.XXXXXX` subdirectory
instead of the previous approach of concatenating XXXXXX to the
file name, which overrided the extension.

From the user's point of view, when opening the attachment, the
filename will have the same as displayed in Evince, which is
what the user likely expects.

https://bugzilla.gnome.org/show_bug.cgi?id=772188

Fixes #711