libdocument: when LOAD_FLAG_NO_CACHE is present, load the cache on demand when needed
The cache is an optimization to avoid taking locks required when using
the backends, but it's also required for some of the features like
getting max/min page sizes or finding a page by its label. This patch
moves the n_pages and info out of what we consider the cache, because
they should always be fast to initialze during the load. The rest of the
cache information is only loaded when actually required when using the
no cache load flag.