ev-sidebar-links: Optimize reverse link lookup for a page
For large documents the linear search for the first link that is on a
certain page is really slow. Because of this scrolling becomes slow
whenever the page changes.

Replace the linear search with a search in a binary tree populated with
the first link on each page and the corresponding GtkTreePath. This way
a specialized binary tree lookup can be used to find the closest
matching link and select that in the treeview.

https://bugzilla.gnome.org/show_bug.cgi?id=779614