shell: Save document to the same path it was opened from
When annotating or filling a form in a document, this must be saved
as a different document as evince does not overwrite documents.
The user can expect to store the modified file in the same place
than the original document, except when the document lives in
a temporary directory (e.g. downloaded automatically with a web
browser), in whose case it must fallback to the Documents
directory (if set) or the the home directory.

Previously, evince assumed the latest directory used, or the place
where an image or attachment was stored last. Such behaviour is
confusing because the latest place opened might have no relation
with the document modified.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=767611