sidebar-thumbnails: keep thumbnails already rendered
Evince renders thumbnails on-the-fly as they get into the
scrolling visible area, but at the same time it will remove
them as they get out of the visible scrolling area, so when
user scrolls back to same position he will notice thumbnails
be recreated.

In pro of a more icing user experience, let's adopt a mixed approach
and keep the thumbnails that the user has already navigated, so
when he scrolls back and forth in the same area no thumbnail
re-generation will be visible. This also matches behaviour with
other pdf readers.

https://bugzilla.gnome.org/show_bug.cgi?id=342110