libview: Make class EvDocumentModel final
This patch makes use of standard GObject macros to declare class
EvDocumentModel.  GObject macros take care of defining a few useful
functions/variables, so the corresponding (so far hand-coded) macros
were removed.  Since the class is not derived from, it is marked "final"
(this restriction can always be lifted at a later date, if needed).  For
that reason GObject will also take care of defining the class structure,
therefore the hand-coded definition was also removed.

That last point also means that an unused class-scoped function pointer
meant to serve as a signal handler (a so-called "class closure") for
signal "page-changed" was removed.  Notice that the signal itself still
exists.

https://bugzilla.gnome.org/show_bug.cgi?id=784914