comics: Check for supported extensions in initial list loop
Instead of looping over all the files once, making a list, then going
over the list once again to check whether those files are supported,
check for support when building the list.

This fixes reporting an error when non-image files are encrypted and
we could read the images just fine, or the separate errors for when the
archive is empty, or there are no supported files in the archive.

https://bugzilla.gnome.org/show_bug.cgi?id=784963