Fix omission from the GtkVBox to GtkBox conversion
Commits 7d6a5f17c and bd9f1a6ad (2013-06-09) that changed the users of
GtkVbox to GtkBox neglected to change the type names in the instance and
class structures (it changed only the GObject type IDs).

Maybe this wasn't noticed before because GtkVBox is in fact derived from
GtkBox, which means that when the GObject machinery accessed the
structure it simply used the smaller prefix memory area corresponding to
a GtkBox.  So this bug may have been inconsequential, but it does
illustrate the amount of trust (read "burden") that GObject places on
the programmer by casting types all over the place (not that it can do
much better while staying in C).

https://bugzilla.gnome.org/show_bug.cgi?id=785783