libdocument: Fixing display of overlapping annotations
If two PDF annotations overlap, one might "shadow" the other,
making it imposible for the user to click on of them if the
smaller one is below the larger one.

This fix calculates the area of bounding box and always puts
the smaller sized annotations on top of the larger ones.

https://bugzilla.gnome.org/show_bug.cgi?id=787070