Allow adding Highlight annotations from text selection
As this is a natural, intuitive way to add text markup
annotations to a document.

The "Highlight Selected Text" action
in the context menu
will show if there is a selection. When
the selection spreads over several pages,
one annotation per page will be added accordingly
to fulfill the user expectation of all selected text
to be annotated.

https://bugzilla.gnome.org/show_bug.cgi?id=763943