Fix add_annotation() to update area based on bounding box
for Text Markup annotations, same way this is done if using
save_annotation().

When pdf_document_annotations_save_annotation() saves a Text Markup
annotation it will, prior to saving, update the area of annotation
to be based on its bounding box. But this was not being done if we
just used pdf_document_annotations_add_annotation() (because we
currently add annotations via a DnD operation that involves both
an initial add_annotation() plus subsequent save_annotation() calls
on motion_notify).

So this fix is needed to be able to correctly create a Text Markup
annotation by a single call to pdf_document_annotations_add_annotation()

Part of https://bugzilla.gnome.org/show_bug.cgi?id=763943