properties: Fix dialog extending past edge of screen
A property with a lot of newlines can extend Nautilus's file properties
dialog past the edge of the screen.  Replace these newlines with spaces
to limit the vertical size and let the label's ellipsizing work
correctly.  Have labels wrap for up to five lines in case someone is
relying on the properties dialog showing large properties.  Place
properties page in a GtkScrolledWindow in case it is still too tall.

https://bugzilla.gnome.org/show_bug.cgi?id=733618