print-operation: Fix orientation of auto-rotate option
When printing a landscape document, if both the "Auto Rotate and Center"
and "Select page size using document page size" options are enabled,
evince will choose the wrong orientation.  This will also occur with any
document if you use a custom paper size which is wider than it is tall
and enable the auto rotate option, but do not enable the "Select page
size" option.

Fix this by having the auto rotate option pick landscape orientation if
and only if the document's orientation does not match the paper's
unrotated orientation.

https://bugzilla.gnome.org/show_bug.cgi?id=792394