comics: Fix reading CBZ files created on macOS
macOS' builtin ZIP file format support will add AppleDouble files to the
archives in order to store information that would otherwise be stored in
extended attributes or resource forks, in a way that's unlikely to be
meddled with by applications that don't know how to handle this
metadata.

But when we store images in such a ZIP file, we'll end up with files
named "__MACOSX/dirname/._page0001.JPG" for example, which our current
code will interpret as being images, and thus pages in our comic.
They're not. As the AppleDouble files always have the "._" prefix,
ignore those.

See https://en.wikipedia.org/wiki/AppleSingle_and_AppleDouble_formats

https://bugzilla.gnome.org/show_bug.cgi?id=792754