shell: Unbind Ctrl-Left and Ctrl-Right from the scrolledwindow
Bug #727529 was introduced when fixing #676040. The rational of
the latter fix is to forward keypress events to the focused
widget before processing Application shortcuts. This allows that
if, for e.g., you Ctrl+left on the search entry, then you jump
between words instead of the default Rotate left action. This is
the intended behavior.

But GtkScrolledWindow binds Ctrl-Left and Ctrl-Right. Hence,
these actions are handled instead of the Rotate left and Rotate right shortcuts defined by the application in the case the horizontal scrollbar is visible.

This is not intended, so unbinding these shortcuts
with CSS allows to Ctrl left and ctrl right to rotate independent of whether
the scrollbars are visible or not.

This solves
https://bugzilla.gnome.org/show_bug.cgi?id=727529