shell: Fix crash in ev_media_player_keys_grab_keys
Asynchronous getting of a DBus proxy can finish after EvMediaPlayerKeys
is finalized but this is not secured by a cancellable and hence it crashes
in callback of the getting of the proxy.
This commit adds the cancellable.

When g_dbus_proxy_new_for_bus() is cancelled because EvMediaPlayerKeys
is finalized, don't use the user_data (which will be invalid) before
checking the return of the call.

The user_data will only be valid if the error returned by the _finish()
call is not G_IO_ERROR_CANCELLED (Bastien Nocera).

https://retrace.fedoraproject.org/faf/reports/1649005/
https://bugzilla.redhat.com/show_bug.cgi?id=1359507

https://bugzilla.gnome.org/show_bug.cgi?id=795978