shell: Move 'Open...' file from document to application
Opening a new document is not an action that applies to the active
document. Therefore, it is a global action which belongs to the
application rather than the window.

We move the visual cue ('menu') from the document window to the
application. Additionally, we remove the toolbar button from
'Recent view' window because it becomes redundant having the visual
action in two places.

However, we still leave available the shortcut (Ctrl-O), which is
done in each window, as it were a global shorcut.

In spite that the action in the UX mainly belongs to the application
rather than a document, we still need to know where to display the
dialog (a parent window), in which screen the document should be
opened, and if the active window is in Recent View mode, use that
window to open the document.

Because we remove the toolbar button in the Recent View, Issue #514
becomes obsolete or duplicated.

Closes #524