libmisc: Let the buttons in search bar get focus
The buttons in the search bar allow the user to navigate through
the previous and next match. For user who rely on keyboard navigation
is important they can select those buttons (other than shortcuts).

The new search entry bar was introduced in commit 16314cd5. There is
no rationale to make the unfocusable. It might have trying to mimic
the behaviour of the old search entry, which also had the issue.

Closes #216