Revert "shell: Move 'Open...' file from document to application"
Because of the new goal (still a proposal, but there is some
consensus that is needed), the Application Menu will be
discouraged or removed. So, it makes sense to revert #524,
and set the goal to fix #514, that is, to use a textual
button in the Recent View, instead of the menu.

This reverts commit 6e0dbaea93f84053cb72e36b2b3b33250004fb94.