shell: Reorganize navigation in the Headerbar
* Remove EvHistoryAction widget from the Headerbar.
  The actions are not always active, and therefore, not a
  primary action that deserves to be always on spot.
* Integrate the history functionality in the
  EvPageActionWidget through a Popover menu.
* Move the navigation menu items from the View menu into
  EvPageActionWidget.

By removing EvHistoryAction widget, there is more space
for the title in the Headerbar, and it looks less
overwhelming.

However, these changes introduces a regression: the
HistoryAction widget had a menu where the user could
select where to jump in the history.

It was a hidden feature, that may or may not be missed.
However, to improve its usability, it could be reintroduced
later as another item in the sidebar, and/or with a keyboard
shortcut (document in the help overlay).