libmisc: Emit "updated" signal at end of search
For long documents (more than 100 pages), EvSearchBox only sends the
"updated" signal every few pages in order to make the GUI more
reponsive, but it can miss the signal for the last few pages of the
search. EvFindSidebar depends on this signal so this can result in
search results for the previous few pages missing from the sidebar.

Fixes https://gitlab.gnome.org/GNOME/evince/issues/917