[pdf] Add support for xmp metadata outside x:xmpmeta element
Based on the specification the *optional* `x:xmpmeta`
element may be placed around the rdf:RDF element.
The element’s namespace URI shall be "adobe:ns:meta/".

Now the functions try to extract the tags from the `rdf:RDF`
element and if is `NULL` they try inside the `x:xmpmeta`.