Move New window from App menu to Window menu
Get rid off the AppMenu.

Remove "Open" from the window because that is a duplicate
action of New window, where there is an action to open a
new document. Additionally, "Open" a document actually
opens a new window, thus having an explicit action to
open a new window, and in that window a new document will
be opened.