annotation: Make annotation title independent of theme
The note text (sticky notes) annotation uses a custom background
color (yellow by default). Having the title adjustable to theme,
breaks dark-themes, because the title becomes white on yellow
background.

We set a fixed black and grey as the foreground colors for the
title, which works on yellow background.

Fixes #664