synctex: Update from version 1.18 to 1.21
Version 1.19 changes the API, normalized the data types, and function
names.

The major change for Evince is to add an extra field in EvSourceLink
and the propotype to create the structure. synctex_display_query
now requires an extra parameters (page-hint) to improve the results,
but the page has not been store before.

Fixes #878