toolbar: Remove bookmarks menu from page selector
After the reorganization of the menu and toolbar, the user can discover
the sidebar, and therefore, they can navigate the document using the
sidebar bookmarks. It is redundant to have the bookmarks also in the
menu, which overpopulate the menu for page selector.

However, we keep the accelerator (shortcut) for Ctrl+D.