libview: Set minimum for Page Forward/Backward scrolling
Page Forward and Backward scrolling leaves a line of context. It jumps
by the height of the view minus the height of a line of text. But if the
text is very tall, scrolling may not work or may scroll in the wrong
direction.

This sets a minimum height for scrolling. If the line is more than 30%
of the page, the height of the line is ignored and it jumps the full
view height.

Fixes #692