fix slow scrolling on last pages of big pdf's
Optimized by not calling compute_border() so many
times in the adjustment-value-changed signal handler.

For that we modified ev_view_get_page_extents() to add
new gboolean parameter saying if we should use the
passed in &border struct as a valid one or otherwise
obtain it as before, calling compute_border().

We replace all occurences of ev_view_get_page_extents()
with the new parameter, setting it to TRUE and passing
a valid &border in the cases it was used inside a loop
iterating over several pages.

Issue #700