djvu: Ignore text regions that are missing text
The DjVu document in issue #948 has text regions that cover the entire
page but only contain an empty string. This makes selecting text with
either the mouse or the Control+A shortcut difficult because the entire
page gets selected. Fix this by ignoring empty strings when determining
the selection region.